dashboard/requirements.txt
streamlit
psycopg2-binary
plotly
watchdog